Separate cube rotation from camera view
In order to draw the initial cube pretty we move the camera
so that it looks at the cube at an angle. But previously we
were changing the actual cube for this, so every time we drew
it, it was rotating at the camera angle.

Now the rotation of the cube happens at the ANGULAR_SPEED only,
and the "camera changes" are stored in a separate temporary
variable.