Add monoclinic and rotate it
Added new crystal system and ability to rotate it

Deleted essential button variables. Now the code is more readable

The rotation axis scaling has been changed. Now the size depends on the figure. This is important because before it was the same for all figures, which meant that for all figures except cubic the axis was smaller than it should be.